提交一下

+++ fair/lib/src/internal/flexbuffer/types.dart
@@ -23,7 +23,9 @@ class BitWidthUtil {
       if (v >> 31 == 0) return BitWidth.width32;
       return BitWidth.width64;
     }
-    return value == _toF32(value) ? BitWidth.width32 : BitWidth.width64;
+    return value == _toF32(value.toDouble())
+        ? BitWidth.width32
+        : BitWidth.width64;
   }
 
   static BitWidth uwidth(num value) {
@@ -34,7 +36,9 @@ class BitWidthUtil {
       if (v >> 32 == 0) return BitWidth.width32;
       return BitWidth.width64;
     }
-    return value == _toF32(value) ? BitWidth.width32 : BitWidth.width64;
+    return value == _toF32(value.toDouble())
+        ? BitWidth.width32
+        : BitWidth.width64;
   }
 
   static BitWidth fromByteWidth(int value) {
